requirements.txt
# AI Cyber Threat Intelligence Collector - Dependencies

# HTTP / API
requests>=2.31.0
aiohttp>=3.9.0
feedparser>=6.0.11

# Data Processing
pydantic>=2.5.0,<3.0.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Storage
aiosqlite>=0.19.0

# Scheduling
apscheduler>=3.10.4

# CLI
click>=8.1.7
rich>=13.7.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
